Changes Complete Laravel 12 with Livewire 3

+++ app/Livewire/Admin/Clients/Index.php
@@ -5,21 +5,26 @@ namespace App\Livewire\Admin\Clients;
 use App\Models\Client;
 use Livewire\Component;
 use Livewire\WithPagination;
+use Livewire\WithoutUrlPagination;
 
 class Index extends Component
 {
-    use WithPagination;
+    use WithPagination, WithoutUrlPagination;
 
     protected $paginationTheme = "bootstrap";
+    public $search = '';
 
     protected $listeners = [
         'done' => 'render'
     ];
 
-    function delete($id)
+    public function delete($id)
     {
         $client = Client::find($id);
-
+        if($client->bookings->count() > 0) {
+            $this->dispatch('done', error: "Cannot delete client with existing bookings. Delete the bookings first.");
+            return;
+        }
         $client->delete();
 
         $this->dispatch('done', success: "Successfully Deleted a Client");
@@ -28,7 +33,10 @@ class Index extends Component
     public function render()
     {
         return view('livewire.admin.clients.index', [
-            'clients' => Client::orderBy('id', 'DESC')->paginate(10)
+            'clients' => Client::where('name', 'like', '%' . $this->search . '%')
+            ->orWhere('email', 'like', '%' . $this->search . '%')
+            ->orderBy('id', 'DESC')
+            ->paginate(10)
         ]);
     }
 }
